+++ .gitignore
@@ -1,1 +1,4 @@
 /target
+/Cargo.lock
+
+**/.env
