+++ Cargo.toml
@@ -16,10 +16,12 @@ path = "src/bin/server.rs"
 
 [dependencies]
 axum = "0.4.5"
-clap = { version = "3.1.0", features = ["derive"] }
-serde = { version = "1.0.136", features = ["derive"] }
-serde_json = "1.0.79"
+clap = { version = "3.1", features = ["derive"] }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+thiserror = "1.0"
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread", "time", "tracing"] }
-tracing = "0.1.30"
-tracing-subscriber = { version = "0.3.8", features = ["env-filter"] }
-tower-http = { version = "0.2.2", features = ["trace"] }
+tracing = "0.1"
+tracing-subscriber = { version = "0.3", features = ["env-filter"] }
+tower = { version = "0.4", features = ["full"] }
+tower-http = { version = "0.2", features = ["full"] }
