+++ src/handler/meter.rs
@@ -1,0 +1,5 @@
+use axum::Router;
+
+pub(crate) fn routes() -> Router {
+    Router::new()
+}
