+++ src/model/health.rs
@@ -1,0 +1,4 @@
+#[derive(Serialize)]
+pub struct Health {
+    pub place_holder: String,
+}
