+++ src/service.rs
@@ -1,0 +1,5 @@
+pub use health::HealthService;
+pub use meter::MeterService;
+
+mod health;
+mod meter;
