+++ src/service/meter.rs
@@ -1,0 +1,3 @@
+pub struct MeterService;
+
+impl MeterService {}
